+++ packages/lumx-react/src/components/button/ButtonRoot.tsx
@@ -35,6 +35,11 @@ interface BaseButtonProps extends GenericProps {
     isSelected?: boolean;
 
     /**
+     * Button disabled state.
+     */
+    isDisabled?: boolean;
+
+    /**
      * Use this property if you specified a URL in the `href` property and you want to customize the link button target property.
      */
     target?: '_self' | '_blank' | '_parent' | '_top';
@@ -122,6 +127,8 @@ const ButtonRoot: React.FC<ButtonRootProps> = (props) => {
     const {
         buttonRef,
         emphasis,
+        disabled,
+        isDisabled = disabled,
         isSelected,
         size,
         color,
@@ -151,6 +158,7 @@ const ButtonRoot: React.FC<ButtonRootProps> = (props) => {
             color: adaptedColor,
             emphasis,
             isSelected,
+            isDisabled,
             prefix: BUTTON_CLASSNAME,
             size,
             theme: emphasis === Emphasis.high && theme,
@@ -161,7 +169,12 @@ const ButtonRoot: React.FC<ButtonRootProps> = (props) => {
 
     if (!isEmpty(props.href)) {
         return (
-            <a {...forwardedProps} ref={buttonRef as RefObject<HTMLAnchorElement>} className={buttonClassName}>
+            <a
+                {...forwardedProps}
+                ref={buttonRef as RefObject<HTMLAnchorElement>}
+                className={buttonClassName}
+                aria-disabled={isDisabled}
+            >
                 {children}
             </a>
         );
@@ -169,6 +182,7 @@ const ButtonRoot: React.FC<ButtonRootProps> = (props) => {
     return (
         <button
             {...forwardedProps}
+            disabled={isDisabled}
             ref={buttonRef as RefObject<HTMLButtonElement>}
             className={buttonClassName}
             name={name}