+++ packages/lumx-react/src/components/checkbox/Checkbox.tsx
@@ -17,7 +17,7 @@ interface CheckboxProps extends GenericProps {
     /** Whether it is checked or not. */
     checked?: boolean;
     /** Is checkbox disabled */
-    disabled?: boolean;
+    isDisabled?: boolean;
     /** Helper */
     helper?: string;
     /** Native input id */
@@ -55,7 +55,6 @@ const CLASSNAME: string = getRootClassName(COMPONENT_NAME);
  * The default value of props.
  */
 const DEFAULT_PROPS: DefaultPropsType = {
-    disabled: false,
     theme: Theme.light,
     checked: false,
 };
@@ -67,7 +66,8 @@ const DEFAULT_PROPS: DefaultPropsType = {
  */
 const Checkbox: React.FC<CheckboxProps> = ({
     className,
-    disabled = DEFAULT_PROPS.disabled,
+    disabled,
+    isDisabled = disabled,
     helper,
     id,
     label,
@@ -93,7 +93,7 @@ const Checkbox: React.FC<CheckboxProps> = ({
                 className,
                 handleBasicClasses({
                     isChecked: checked,
-                    isDisabled: disabled,
+                    isDisabled,
                     isUnchecked: !checked,
                     prefix: CLASSNAME,
                     theme,
@@ -106,7 +106,7 @@ const Checkbox: React.FC<CheckboxProps> = ({
                     type="checkbox"
                     id={inputId}
                     className={`${CLASSNAME}__input-native`}
-                    tabIndex={disabled ? -1 : 0}
+                    tabIndex={isDisabled ? -1 : 0}
                     name={name}
                     value={value}
                     checked={checked}