+++ packages/lumx-react/src/components/switch/Switch.tsx
@@ -22,6 +22,9 @@ interface SwitchProps extends GenericProps {
     /** Whether it is checked or not. */
     checked?: boolean;
 
+    /** Switch disabled state. */
+    isDisabled?: boolean;
+
     /**
      * A small help to display below.
      */
@@ -83,6 +86,8 @@ const Switch: React.FC<SwitchProps> = ({
     className,
     children,
     checked = DEFAULT_PROPS.checked,
+    disabled,
+    isDisabled = disabled,
     helper,
     name,
     onChange,
@@ -107,13 +112,14 @@ const Switch: React.FC<SwitchProps> = ({
                 handleBasicClasses({
                     prefix: CLASSNAME,
                     checked: Boolean(checked),
-                    disabled: forwardedProps.disabled,
+                    isDisabled,
                     position,
                     theme,
                     unchecked: !Boolean(checked),
                 }),
                 { [`${CSS_PREFIX}-custom-colors`]: useCustomColors },
             )}
+            aria-disabled={isDisabled}
         >
             <div className={`${CLASSNAME}__input-wrapper`}>
                 <input
@@ -122,6 +128,7 @@ const Switch: React.FC<SwitchProps> = ({
                     className={`${CLASSNAME}__input-native`}
                     name={name}
                     value={value}
+                    disabled={isDisabled}
                     checked={checked}
                     onChange={handleChange}
                 />