+++ packages/lumx-react/src/components/table/TableRow.tsx
@@ -24,11 +24,6 @@ interface TableRowProps extends GenericProps {
 }
 
 /**
- * Define the types of the default props.
- */
-interface DefaultPropsType extends Partial<TableRowProps> {}
-
-/**
  * The display name of the component.
  */
 const COMPONENT_NAME = `${COMPONENT_PREFIX}TableRow`;
@@ -39,15 +34,6 @@ const COMPONENT_NAME = `${COMPONENT_PREFIX}TableRow`;
 const CLASSNAME: string = getRootClassName(COMPONENT_NAME, true);
 
 /**
- * The default value of props.
- */
-const DEFAULT_PROPS: DefaultPropsType = {
-    isClickable: false,
-    isDisabled: false,
-    isSelected: false,
-};
-
-/**
  * The TableRow component displays an HTML Table Row, which contains table cells.
  *
  * @return The component.
@@ -55,9 +41,10 @@ const DEFAULT_PROPS: DefaultPropsType = {
 const TableRow: React.FC<TableRowProps> = ({
     children,
     className,
-    isClickable = DEFAULT_PROPS.isClickable,
-    isDisabled = DEFAULT_PROPS.isDisabled,
-    isSelected = DEFAULT_PROPS.isSelected,
+    isClickable,
+    disabled,
+    isDisabled = disabled,
+    isSelected,
     ...forwardedProps
 }) => (
     <tr
@@ -72,6 +59,7 @@ const TableRow: React.FC<TableRowProps> = ({
             }),
         )}
         tabIndex={isClickable && !isDisabled ? 0 : -1}
+        aria-disabled={isDisabled}
     >
         {children}
     </tr>
@@ -79,4 +67,4 @@ const TableRow: React.FC<TableRowProps> = ({
 
 TableRow.displayName = COMPONENT_NAME;
 
-export { CLASSNAME, DEFAULT_PROPS, TableRow, TableRowProps };
+export { CLASSNAME, TableRow, TableRowProps };