+++ packages/site-demo/content/product/components/button/react/emphasis-low.tsx
@@ -10,7 +10,7 @@ const App = ({ theme }: any) => (
             Default
         </Button>
 
-        <Button disabled emphasis={Emphasis.low} theme={theme}>
+        <Button isDisabled emphasis={Emphasis.low} theme={theme}>
             Disabled
         </Button>