+++ packages/site-demo/content/product/components/button/react/small.tsx
@@ -11,7 +11,7 @@ const App = ({ theme }: any) => (
                 Default
             </Button>
 
-            <Button disabled size={Size.s} theme={theme}>
+            <Button isDisabled size={Size.s} theme={theme}>
                 Disabled
             </Button>
 
@@ -38,7 +38,7 @@ const App = ({ theme }: any) => (
                 Default
             </Button>
 
-            <Button disabled emphasis={Emphasis.medium} size={Size.s} theme={theme}>
+            <Button isDisabled emphasis={Emphasis.medium} size={Size.s} theme={theme}>
                 Disabled
             </Button>
             <Button emphasis={Emphasis.medium} leftIcon={mdiPencil} size={Size.s} theme={theme}>
@@ -64,7 +64,7 @@ const App = ({ theme }: any) => (
                 Default
             </Button>
 
-            <Button disabled emphasis={Emphasis.low} size={Size.s} theme={theme}>
+            <Button isDisabled emphasis={Emphasis.low} size={Size.s} theme={theme}>
                 Disabled
             </Button>