+++ packages/site-demo/content/product/components/radio-button/react/helper.tsx
@@ -20,7 +20,7 @@ const App = ({ theme }: any) => {
                 />
 
                 <RadioButton
-                    disabled
+                    isDisabled
                     helper="Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed posuere faucibus efficitur."
                     label="Radio button with help 2"
                     name="test2"